svg: Move svg dom support under modules/svg/ and separate basic SVG support (canvas) from SVG rendering support
