+++ skia-safe/src/interop/stream.rs
@@ -7,7 +7,10 @@ use crate::{prelude::*, Data};
 use skia_bindings::{
     self as sb, SkDynamicMemoryWStream, SkMemoryStream, SkStream, SkStreamAsset, SkWStream,
 };
-use std::{ffi, fmt, io, marker::PhantomData, ptr};
+use std::{fmt, marker::PhantomData, ptr};
+
+#[cfg(feature = "svg")]
+use std::{ffi, io};
 
 /// Trait representing an Skia allocated Stream type with a base class of SkStream.
 #[repr(transparent)]
@@ -164,23 +167,28 @@ impl DynamicMemoryWStream {
     }
 }
 
+#[cfg(feature = "svg")]
 pub struct RustStream<'a> {
     inner: Handle<sb::RustStream>,
     _phantom: PhantomData<&'a mut ()>,
 }
 
+#[cfg(feature = "svg")]
 impl RustStream<'_> {
     pub fn stream_mut(&mut self) -> &mut SkStream {
         self.inner.native_mut().base_mut()
     }
 }
 
+#[cfg(feature = "svg")]
 impl NativeBase<SkStream> for sb::RustStream {}
 
+#[cfg(feature = "svg")]
 impl NativeDrop for sb::RustStream {
     fn drop(&mut self) {}
 }
 
+#[cfg(feature = "svg")]
 impl<'a> RustStream<'a> {
     pub fn new<T: io::Read>(val: &'a mut T) -> Self {
         unsafe extern "C" fn read_trampoline<T>(
